refactor annotation scanners
