+++ src/Adamgoose/Routing/Annotations/Scanner.php
@@ -1,45 +1,17 @@
 <?php namespace Adamgoose\Routing\Annotations;
 
 use ReflectionClass;
-use Symfony\Component\Finder\Finder;
-use Doctrine\Common\Annotations\AnnotationRegistry;
 use Doctrine\Common\Annotations\SimpleAnnotationReader;
+use Adamgoose\AnnotationScanner;
 
-class Scanner {
-
-	/**
-	 * The path to scan for annotations.
-	 *
-	 * @var array
-	 */
-	protected $scan;
+class Scanner extends AnnotationScanner {
 
 	/**
-	 * Create a new scanner instance.
+	 * The namespace to check for annotation reader annotations
 	 *
-	 * @param  array  $scan
-	 * @return void
+	 * @var string
 	 */
-	public function __construct(array $scan)
-	{
-		$this->scan = $scan;
-
-		foreach (Finder::create()->files()->in(__DIR__.'/Annotations') as $file)
-		{
-			AnnotationRegistry::registerFile($file->getRealPath());
-		}
-	}
-
-	/**
-	 * Create a new scanner instance.
-	 *
-	 * @param  array  $scan
-	 * @return static
-	 */
-	public static function create(array $scan)
-	{
-		return new static($scan);
-	}
+	protected $namespace = 'Adamgoose\Routing\Annotations\Annotations';
 
 	/**
 	 * Convert the scanned annotations into route definitions.
@@ -117,42 +89,4 @@ class Scanner {
 		foreach ($annotations as $annotation)
 			$annotation->modify($endpoint, $class->getMethod($method));
 	}
-
-	/**
-	 * Get all of the ReflectionClass instances in the scan array.
-	 *
-	 * @return array
-	 */
-	protected function getClassesToScan()
-	{
-		$classes = [];
-
-		foreach ($this->scan as $scan)
-		{
-			try
-			{
-				$classes[] = new ReflectionClass($scan);
-			}
-			catch (Exception $e)
-			{
-				//
-			}
-		}
-
-		return $classes;
-	}
-
-	/**
-	 * Get an annotation reader instance.
-	 *
-	 * @return \Doctrine\Common\Annotations\SimpleAnnotationReader
-	 */
-	protected function getReader()
-	{
-		with($reader = new SimpleAnnotationReader)
-				->addNamespace('Adamgoose\Routing\Annotations\Annotations');
-
-		return $reader;
-	}
-
 }
